+++ src/utils/RoleDefinitions.tsx
@@ -24,8 +24,8 @@ export class RoleDefinitions {
         return this.userIsManager(roles);
     }
 
-    static userCanEditNotes(roles?: RoleType[]): boolean {
-        return this.userIsManager(roles);
+    static userCanAddNotes(request: IRequirementsRequest, roles?: RoleType[], currentUser?: IPerson): boolean {
+        return this.userIsManager(roles) || currentUser?.Id === request.Requester.Id || currentUser?.Id === request.Approver.Id || currentUser?.Id === request.Author.Id;
     }
 
     static userCanChangeStatus(request: IRequirementsRequest, newStatus: RequestStatuses | null, currentUser?: IPerson, roles?: RoleType[]): boolean {